+++ src/pitloom/__about__.py
@@ -4,4 +4,4 @@
 
 """Package information for Pitloom."""
 
-__version__ = "0.4.1"
+__version__ = "0.5.0"