+++ CHANGELOG.md
@@ -15,7 +15,17 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Full release notes: <https://github.com/bact/pitloom/releases>
-- Commit history: <https://github.com/bact/pitloom/compare/v0.4.0...v0.4.1>
+- Commit history: <https://github.com/bact/pitloom/compare/v0.4.1...v0.5.0>
+
+## [0.5.0] - 2026-04-29
+
+### Added
+
+- Support projects build with setuptools ([#59])
+- Detect SPDX License ID from license text ([#60])
+
+[#59]: https://github.com/bact/pitloom/pull/59
+[#60]: https://github.com/bact/pitloom/pull/60
 
 ## [0.4.1] - 2026-04-02
 
@@ -99,6 +109,7 @@ release because "Loom" and "Pyloom" were unavailable on PyPI.
 
 ---
 
+[0.5.0]: https://github.com/bact/pitloom/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/bact/pitloom/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/bact/pitloom/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/bact/pitloom/compare/v0.2.0...v0.3.0